Changed JSON format for CarCommunicator→CarAPI and specified URI for CarDataStream→CarCommunicator
The JSON format for communication between CarCommunicator and CarAPI now matches the format for CarDataStream→CarCommunicator. This is done to remove the timestamp from PI to server, since it is deemed unnecessary. 

The URL for requests from CarDataStream to CarCommunicator has also been specified, since it was left out in earlier revisions.